+++ unchecked-xteps/src/main/java/com/plugatar/xteps/unchecked/stepobject/ConsumerStep.java
@@ -20,7 +20,7 @@ import com.plugatar.xteps.base.XtepsException;
 import com.plugatar.xteps.unchecked.UncheckedXteps;
 
 import static com.plugatar.xteps.unchecked.stepobject.StepObjectsUtils.humanReadableStepNameOfClass;
-import static com.plugatar.xteps.unchecked.stepobject.StepObjectsUtils.stepNameWithPrefix;
+import static com.plugatar.xteps.unchecked.stepobject.StepObjectsUtils.stepNameWithKeyword;
 
 /**
  * Consumer step. This step will be executed and reported when calling the {@link #accept(Object)} method.
@@ -98,14 +98,14 @@ public class ConsumerStep<T> implements ThrowingConsumer<T, RuntimeException> {
     }
 
     /**
-     * Returns a new ConsumerStep with given prefix in the step name.
+     * Returns a new ConsumerStep with given keyword in the step name.
      *
-     * @param stepNamePrefix the step name prefix
-     * @return ConsumerStep with given prefix in the step name
+     * @param keyword the keyword
+     * @return ConsumerStep with given keyword in the step name
      */
-    public final ConsumerStep<T> withNamePrefix(final String stepNamePrefix) {
+    public final ConsumerStep<T> withKeyword(final String keyword) {
         return new ConsumerStep<>(
-            stepNameWithPrefix(stepNamePrefix, this.stepName), this.stepDescription, this.step
+            stepNameWithKeyword(keyword, this.stepName), this.stepDescription, this.step
         );
     }