+++ xteps/src/main/java/com/plugatar/xteps/checked/chain/base/Base1CtxSC.java
@@ -101,19 +101,19 @@ public interface Base1CtxSC<C> {
     ) throws E;
 
     /**
-     * Performs and reports given step with given prefix in the step name and returns
+     * Performs and reports given step with given keyword in the step name and returns
      * this steps chain.
      *
-     * @param stepNamePrefix the step name prefix
-     * @param step           the step
-     * @param <E>            the {@code step} exception type
+     * @param keyword the keyword
+     * @param step    the step
+     * @param <E>     the {@code step} exception type
      * @return this steps chain
-     * @throws XtepsException if {@code stepNamePrefix} or {@code step} is null
+     * @throws XtepsException if {@code keyword} or {@code step} is null
      *                        or if it's impossible to correctly report the step
      * @throws E              if {@code step} threw exception
      */
     <E extends Throwable> BaseCtxSC<?> step(
-        String stepNamePrefix,
+        String keyword,
         ConsumerStep<? super C, ? extends E> step
     ) throws E;
 
@@ -167,20 +167,20 @@ public interface Base1CtxSC<C> {
     ) throws E;
 
     /**
-     * Performs and reports given step with given prefix in the step name and returns
+     * Performs and reports given step with given keyword in the step name and returns
      * a contextual steps chain of the new context.
      *
-     * @param stepNamePrefix the step name prefix
-     * @param step           the step
-     * @param <U>            the context type
-     * @param <E>            the {@code step} exception type
+     * @param keyword the keyword
+     * @param step    the step
+     * @param <U>     the context type
+     * @param <E>     the {@code step} exception type
      * @return contextual steps chain
-     * @throws XtepsException if {@code stepNamePrefix} or {@code step} is null
+     * @throws XtepsException if {@code keyword} or {@code step} is null
      *                        or if it's impossible to correctly report the step
      * @throws E              if {@code step} threw exception
      */
     <U, E extends Throwable> BaseCtxSC<?> stepToContext(
-        String stepNamePrefix,
+        String keyword,
         FunctionStep<? super C, ? extends U, ? extends E> step
     ) throws E;
 
@@ -238,20 +238,20 @@ public interface Base1CtxSC<C> {
     ) throws E;
 
     /**
-     * Performs and reports given step with given prefix in the step name and returns
+     * Performs and reports given step with given keyword in the step name and returns
      * the step result.
      *
-     * @param stepNamePrefix the step name prefix
-     * @param step           the step
-     * @param <R>            the result type
-     * @param <E>            the {@code step} exception type
+     * @param keyword the keyword
+     * @param step    the step
+     * @param <R>     the result type
+     * @param <E>     the {@code step} exception type
      * @return {@code step} result
-     * @throws XtepsException if {@code stepNamePrefix} or {@code step} is null
+     * @throws XtepsException if {@code keyword} or {@code step} is null
      *                        or if it's impossible to correctly report the step
      * @throws E              if {@code step} threw exception
      */
     <R, E extends Throwable> R stepTo(
-        String stepNamePrefix,
+        String keyword,
         FunctionStep<? super C, ? extends R, ? extends E> step
     ) throws E;