+++ xteps/src/main/java/com/plugatar/xteps/checked/chain/impl/Mem2CtxSCImpl.java
@@ -243,13 +243,13 @@ public class Mem2CtxSCImpl<C, C2, PS extends BaseCtxSC<?>> implements Mem2CtxSC<
 
     @Override
     public final <E extends Throwable> Mem2CtxSC<C, C2, PS> step(
-        final String stepNamePrefix,
+        final String keyword,
         final RunnableStep<? extends E> step
     ) throws E {
-        if (stepNamePrefix == null) { this.throwNullArgException("stepNamePrefix"); }
+        if (keyword == null) { this.throwNullArgException("keyword"); }
         if (step == null) { this.throwNullArgException("step"); }
         this.execAction(() -> {
-            step.withNamePrefix(stepNamePrefix).run();
+            step.withKeyword(keyword).run();
             return null;
         });
         return this;
@@ -257,13 +257,13 @@ public class Mem2CtxSCImpl<C, C2, PS extends BaseCtxSC<?>> implements Mem2CtxSC<
 
     @Override
     public final <E extends Throwable> Mem2CtxSC<C, C2, PS> step(
-        final String stepNamePrefix,
+        final String keyword,
         final ConsumerStep<? super C, ? extends E> step
     ) throws E {
-        if (stepNamePrefix == null) { this.throwNullArgException("stepNamePrefix"); }
+        if (keyword == null) { this.throwNullArgException("keyword"); }
         if (step == null) { this.throwNullArgException("step"); }
         this.execAction(() -> {
-            step.withNamePrefix(stepNamePrefix).accept(this.context);
+            step.withKeyword(keyword).accept(this.context);
             return null;
         });
         return this;
@@ -271,13 +271,13 @@ public class Mem2CtxSCImpl<C, C2, PS extends BaseCtxSC<?>> implements Mem2CtxSC<
 
     @Override
     public final <E extends Throwable> Mem2CtxSC<C, C2, PS> step(
-        final String stepNamePrefix,
+        final String keyword,
         final BiConsumerStep<? super C, ? super C2, ? extends E> step
     ) throws E {
-        if (stepNamePrefix == null) { this.throwNullArgException("stepNamePrefix"); }
+        if (keyword == null) { this.throwNullArgException("keyword"); }
         if (step == null) { this.throwNullArgException("step"); }
         this.execAction(() -> {
-            step.withNamePrefix(stepNamePrefix).accept(this.context, this.context2);
+            step.withKeyword(keyword).accept(this.context, this.context2);
             return null;
         });
         return this;
@@ -397,34 +397,34 @@ public class Mem2CtxSCImpl<C, C2, PS extends BaseCtxSC<?>> implements Mem2CtxSC<
 
     @Override
     public final <U, E extends Throwable> Mem3CtxSC<U, C, C2, Mem2CtxSC<C, C2, PS>> stepToContext(
-        final String stepNamePrefix,
+        final String keyword,
         final SupplierStep<? extends U, ? extends E> step
     ) throws E {
-        if (stepNamePrefix == null) { this.throwNullArgException("stepNamePrefix"); }
+        if (keyword == null) { this.throwNullArgException("keyword"); }
         if (step == null) { this.throwNullArgException("step"); }
-        return this.newMem2CtxStepsChain(this.execAction(step.withNamePrefix(stepNamePrefix)));
+        return this.newMem2CtxStepsChain(this.execAction(step.withKeyword(keyword)));
     }
 
     @Override
     public final <U, E extends Throwable> Mem3CtxSC<U, C, C2, Mem2CtxSC<C, C2, PS>> stepToContext(
-        final String stepNamePrefix,
+        final String keyword,
         final FunctionStep<? super C, ? extends U, ? extends E> step
     ) throws E {
-        if (stepNamePrefix == null) { this.throwNullArgException("stepNamePrefix"); }
+        if (keyword == null) { this.throwNullArgException("keyword"); }
         if (step == null) { this.throwNullArgException("step"); }
         return this.newMem2CtxStepsChain(this.execAction(
-            () -> step.withNamePrefix(stepNamePrefix).apply(this.context)));
+            () -> step.withKeyword(keyword).apply(this.context)));
     }
 
     @Override
     public final <U, E extends Throwable> Mem3CtxSC<U, C, C2, Mem2CtxSC<C, C2, PS>> stepToContext(
-        final String stepNamePrefix,
+        final String keyword,
         final BiFunctionStep<? super C, ? super C2, ? extends U, ? extends E> step
     ) throws E {
-        if (stepNamePrefix == null) { this.throwNullArgException("stepNamePrefix"); }
+        if (keyword == null) { this.throwNullArgException("keyword"); }
         if (step == null) { this.throwNullArgException("step"); }
         return this.newMem2CtxStepsChain(this.execAction(
-            () -> step.withNamePrefix(stepNamePrefix).apply(this.context, this.context2)));
+            () -> step.withKeyword(keyword).apply(this.context, this.context2)));
     }
 
     @Override
@@ -514,32 +514,32 @@ public class Mem2CtxSCImpl<C, C2, PS extends BaseCtxSC<?>> implements Mem2CtxSC<
 
     @Override
     public final <R, E extends Throwable> R stepTo(
-        final String stepNamePrefix,
+        final String keyword,
         final SupplierStep<? extends R, ? extends E> step
     ) throws E {
-        if (stepNamePrefix == null) { this.throwNullArgException("stepNamePrefix"); }
+        if (keyword == null) { this.throwNullArgException("keyword"); }
         if (step == null) { this.throwNullArgException("step"); }
-        return this.execAction(step.withNamePrefix(stepNamePrefix));
+        return this.execAction(step.withKeyword(keyword));
     }
 
     @Override
     public final <R, E extends Throwable> R stepTo(
-        final String stepNamePrefix,
+        final String keyword,
         final FunctionStep<? super C, ? extends R, ? extends E> step
     ) throws E {
-        if (stepNamePrefix == null) { this.throwNullArgException("stepNamePrefix"); }
+        if (keyword == null) { this.throwNullArgException("keyword"); }
         if (step == null) { this.throwNullArgException("step"); }
-        return this.execAction(() -> step.withNamePrefix(stepNamePrefix).apply(this.context));
+        return this.execAction(() -> step.withKeyword(keyword).apply(this.context));
     }
 
     @Override
     public final <R, E extends Throwable> R stepTo(
-        final String stepNamePrefix,
+        final String keyword,
         final BiFunctionStep<? super C, ? super C2, ? extends R, ? extends E> step
     ) throws E {
-        if (stepNamePrefix == null) { this.throwNullArgException("stepNamePrefix"); }
+        if (keyword == null) { this.throwNullArgException("keyword"); }
         if (step == null) { this.throwNullArgException("step"); }
-        return this.execAction(() -> step.withNamePrefix(stepNamePrefix).apply(this.context, this.context2));
+        return this.execAction(() -> step.withKeyword(keyword).apply(this.context, this.context2));
     }
 
     @Override