+++ xteps/src/main/java/com/plugatar/xteps/checked/stepobject/RunnableStep.java
@@ -20,7 +20,7 @@ import com.plugatar.xteps.base.XtepsException;
 import com.plugatar.xteps.checked.Xteps;
 
 import static com.plugatar.xteps.checked.stepobject.StepObjectsUtils.humanReadableStepNameOfClass;
-import static com.plugatar.xteps.checked.stepobject.StepObjectsUtils.stepNameWithPrefix;
+import static com.plugatar.xteps.checked.stepobject.StepObjectsUtils.stepNameWithKeyword;
 
 /**
  * Runnable step. This step will be executed and reported when calling the {@link #run()} method.
@@ -96,14 +96,14 @@ public class RunnableStep<E extends Throwable> implements ThrowingRunnable<E> {
     }
 
     /**
-     * Returns a new RunnableStep with given prefix in the step name.
+     * Returns a new RunnableStep with given keyword in the step name.
      *
-     * @param stepNamePrefix the step name prefix
-     * @return RunnableStep with given prefix in the step name
+     * @param keyword the keyword
+     * @return RunnableStep with given keyword in the step name
      */
-    public final RunnableStep<E> withNamePrefix(final String stepNamePrefix) {
+    public final RunnableStep<E> withKeyword(final String keyword) {
         return new RunnableStep<>(
-            stepNameWithPrefix(stepNamePrefix, this.stepName), this.stepDescription, this.step
+            stepNameWithKeyword(keyword, this.stepName), this.stepDescription, this.step
         );
     }